Move inner class from `BuildOption` to `Option`.
