+++ subprojects/build-option/src/main/java/org/gradle/internal/buildoption/Option.java
@@ -21,4 +21,43 @@ package org.gradle.internal.buildoption;
  * property. Some options may also be configurable via the Gradle API, for example feature previews.
  */
 public interface Option {
+    abstract class Value<T> {
+        public abstract boolean isExplicit();
+
+        public abstract T get();
+
+        /**
+         * Creates the default value for an option.
+         */
+        public static <T> Value<T> defaultValue(final T value) {
+            return new Value<T>() {
+                @Override
+                public boolean isExplicit() {
+                    return false;
+                }
+
+                @Override
+                public T get() {
+                    return value;
+                }
+            };
+        }
+
+        /**
+         * Creates an explicit value for an option.
+         */
+        public static <T> Value<T> value(final T value) {
+            return new Value<T>() {
+                @Override
+                public boolean isExplicit() {
+                    return true;
+                }
+
+                @Override
+                public T get() {
+                    return value;
+                }
+            };
+        }
+    }
 }
